% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_services.R
\docType{data}
\name{baseApi}
\alias{baseApi}
\title{pingToServer}
\format{
An object of class \code{character} of length 1.
}
\usage{
baseApi
}
\value{
Returns a data.frame as result.
}
\description{
pingToServer
}
\keyword{datasets}
