% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_services.R
\name{fetchSymbolAggTrades}
\alias{fetchSymbolAggTrades}
\title{fetchSymbolAggTrades}
\usage{
fetchSymbolAggTrades(symbol, limit)
}
\arguments{
\item{symbol}{ticker pair}

\item{limit}{limits of json data}
}
\value{
Returns a data.frame as result.
}
\description{
fetchSymbolAggTrades
}
