% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_services.R
\name{fetchSymbolOrderBook}
\alias{fetchSymbolOrderBook}
\title{fetchSymbolOrderBook}
\usage{
fetchSymbolOrderBook(symbol, limit = 2)
}
\value{
Returns a data.frame as result contains order book details.
}
\description{
fetchSymbolOrderBook
}
