% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_services.R
\name{fetchSymbolTrades}
\alias{fetchSymbolTrades}
\title{fetchSymbolTrades}
\usage{
fetchSymbolTrades(symbol, limit)
}
\value{
Returns a data.frame as result contains symbol trades details.
}
\description{
fetchSymbolTrades
}
