% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_services.R
\name{fetchTicker24hPrice}
\alias{fetchTicker24hPrice}
\title{fetchTicker24hPrice}
\usage{
fetchTicker24hPrice(symbol)
}
\value{
Returns a data.frame as result contains ticker 24 hours price details.
}
\description{
fetchTicker24hPrice
}
