% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_services.R
\name{fetchTickerPrice}
\alias{fetchTickerPrice}
\title{fetchTicker24hPrice}
\usage{
fetchTickerPrice(symbol)
}
\value{
Returns a data.frame as result contains ticker price details.
}
\description{
fetchTicker24hPrice
}
