% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_services.R
\name{fetchSymbolAggTrades}
\alias{fetchSymbolAggTrades}
\title{fetchSymbolAggTrades}
\usage{
fetchSymbolAggTrades(symbol, limit)
}
\value{
Returns a data.frame as result contains symbol aggregate trades details.
}
\description{
fetchSymbolAggTrades
}
