% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_services.R
\name{fetchSymbolAveragePrice}
\alias{fetchSymbolAveragePrice}
\title{fetchSymbolAveragePrice}
\usage{
fetchSymbolAveragePrice(symbol)
}
\value{
Returns a data.frame as result.
}
\description{
fetchSymbolAveragePrice
}
