% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_services.R
\name{fetchTickerPrice}
\alias{fetchTickerPrice}
\title{fetchTickerPrice}
\usage{
fetchTickerPrice(symbol)
}
\arguments{
\item{symbol}{ticker pair}
}
\value{
Returns a data.frame as result.
}
\description{
fetchTickerPrice
}
